+++ packages/flow/test/Flow.test.ts
@@ -180,6 +180,7 @@ const makeFakeGitSkippingFirstDiff = (
 const failingHosting: GitHubToolShape = {
   createPr: (_title, _body, _base, _draft) => unusedFlow,
   readIssue: (_ref) => unusedFlow,
+  readIssueComments: (_ref) => unusedFlow,
   writeIssueComment: (_ref, _body) => unusedFlow,
   editIssueComment: (_comment, _body) => unusedFlow,
   writePrComment: (_pr, _body) => unusedFlow,